find by categories and other api
We can use these api to call by categories with extra filters.
ALso we can use suggested keywords when user clicks on the search
button. It will automatically suggest a keyword.
So we can display keyword .. for eg. hairy potter shows suggestion as
harry potter. We can dispay this suggestion on top of deals and we can
click on it